move signing token to centjs
