+++ centrifuge-js/package.json
@@ -28,11 +28,13 @@
   },
   "dependencies": {
     "@polkadot/api": "^8.0.1",
+    "@polkadot/keyring": "^10.1.7",
     "@polkadot/util-crypto": "^9.0.1",
     "bn.js": "^5.2.0",
     "clp-wasm": "^0.0.15",
     "decimal.js-light": "^2.5.1",
     "isomorphic-fetch": "^3.0.0",
+    "jw3t": "^1.0.8",
     "rxjs": "^7.4.0"
   },
   "devDependencies": {
