+++ centrifuge-js/src/modules/pod.ts
@@ -1,5 +1,8 @@
+import { Signer } from '@polkadot/api/types'
+import { Keyring } from '@polkadot/keyring'
 import { u8aToHex } from '@polkadot/util'
 import { decodeAddress } from '@polkadot/util-crypto'
+import * as jw3t from 'jw3t'
 
 type JobResponse = {
   JobID: string
@@ -106,6 +109,39 @@ export function getPodModule() {
     return res as T
   }
 
+  async function signToken(args: [address: string, onBehalfOf: string, signer: Signer]) {
+    const [address, onBehalfOf, signer] = args
+    const header = {
+      algorithm: 'sr25519',
+      token_type: 'JW3T',
+      address_type: 'ss58',
+    }
+    const now = Math.floor(Date.now() / 1000)
+    const payload = {
+      address,
+      on_behalf_of: onBehalfOf,
+      proxy_type: 'pod_auth',
+      expires_at: String(now + 60 * 60 * 24),
+      issued_at: String(now),
+      not_before: String(now),
+    }
+    const content = new jw3t.JW3TContent(header, payload)
+
+    const keyring = new Keyring({ type: 'sr25519' })
+    const account = keyring.addFromAddress(address)
+
+    const polkaJsSigner = new jw3t.PolkaJsSigner({
+      account,
+      // @ts-expect-error Signer type version mismatch
+      signer,
+    })
+    const jw3tSigner = new jw3t.JW3TSigner(polkaJsSigner, content)
+    const { base64Content, base64Sig } = await jw3tSigner.getSignature()
+    const token = `${base64Content}.${base64Sig}`
+
+    return { payload, token }
+  }
+
   async function getJob(args: [podUrl: string, token: string, jobId: string]) {
     const [podUrl, token, jobId] = args
     const res = await callPod<JobResponse>(podUrl, `v2/jobs/${jobId}`, 'get', token)
@@ -187,6 +223,7 @@ export function getPodModule() {
   }
 
   return {
+    signToken,
     getJob,
     getAccount,
     createDocument,
